+++ src/utils/talespire.js
@@ -1,3 +1,4 @@
+import { doubleDiceExpression } from "~/utils/diceUtils";
 import { settings } from "~/utils/storage";
 
 const modifierAction = (action, name) => {
@@ -8,6 +9,8 @@ const modifierAction = (action, name) => {
       return { name: `${name} (DIS)`, extraDice: true };
     case "adv-dis":
       return { name: `${name} (ADV/DIS)`, extraDice: true };
+    case "crit":
+      return { name: `${name} (CRIT)`, extraDice: false };
 
     default:
       return { name, extraDice: false };
@@ -67,6 +70,10 @@ export const talespireLink = (elem, label, dice, diceLabel) => {
     event.stopPropagation();
 
     const { name, extraDice } = checkModifierKeys(event, label);
+
+    if (name.includes("CRIT")) {
+      dice = doubleDiceExpression(dice);
+    }
 
     triggerTalespire(name, dice, extraDice);
   });